scripts/apify_tiktok_scraper/requirements.txt
apify-client>=1.8.1
python-dotenv>=1.0.1
